undo one of those cleanups we want the integer rounding behavior here

+++ Code/Geometry/UniformGrid3D.cpp
@@ -154,8 +154,8 @@ Point3D UniformGrid3D::getGridPointLoc(unsigned int pointId) const {
   }
   Point3D res;
   res.x = (pointId % d_numX) * d_spacing;
-  res.y = ((double)(pointId % (d_numX * d_numY)) / d_numX) * d_spacing;
-  res.z = ((double)pointId / (d_numX * d_numY)) * d_spacing;
+  res.y = ((pointId % (d_numX * d_numY)) / d_numX) * d_spacing;
+  res.z = (pointId / (d_numX * d_numY)) * d_spacing;
   res += d_offSet;  // d_origin;
   return res;
 }
